MAINT: move to hyperu to xsf/cpu and add float overload
